Updated SeeLevel dBus service polling frequency to avoid missing updates
The SeeLevel N2K system often sends updates very close together. These update the SeeLevel dBus object which is polled my the repeater process. CheckSeeLevel () is the method that polls the dBus object. GetValue () calls take a significant amount of time so the code now checks to see if a new tank is being reported before processing the poll further. In addition, repeater services are updated only every 1 second.

Before these changes, a tank would indicate "NO RESPONSE" for a long period of time, possibly requiring reset of the CCGX and/or SeeLevel.

These changes are not perfect. Timeouts (missing a tank for 10 seconds or more) still occur -- about 0.8%. The good news is that the "NO REPSONSE" indications last only 1-2 seconds.