added back second read of SeeLevel's values
When reading SeeLevel values, the tank number can change, invalidating the other parameters. Reading the value set a second time reduces the read failure significantly. The last change removed this second read. This change puts it back.

Also a note: while I'm debugging this code, I'm using Hardware Version to report the total number of seconds I detect a timeout for each tank. I use this value because it is shown on the GUI.