Trap invalid signal receiver values
If the SeeLevel dBus service goes off line because the SeeLevel device stops responding, dBus exceptions occurred when attempting to use the change values passed in the signal handler. A test for a valid number is no made in all three signal handlers before attempting to convert a nonexistent value to a number.