Version-independent installer / Enhanced Mobile Overview page
Previous versions of the installer relied on different OverviewMobile.qml files for variouls Venus versions. In reality, the necessary changes were the same for all versions, so the origial file is now modified to include the code changes. The same logic now applies to all versions, and hopefully future ones as well. sed (Stream EDitor) is used to modify the original qml file.

A while back, I reworked the Mobile Overview page to reorganize the tiles for a left to right power flow (mostly) and added voltage, current and frequency to the AC input and output tiles. The System and Status tiles were combined. The ESS reason indicators were changed from #1, etc to a text message. These messages were merged with the notications into a single scrolling marquee.

The installer now presents an option of modifying the original Mobile Overview page or replacing it with the enhanced version described above.

The Tank Tile modificaitons were also made optional during the install process.